requirements.txt
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.65.0
openai>=1.0.0
python-dotenv>=1.0.0
oracledb>=2.0.0
